backend/pyproject.toml
[project]
name = "donnaai-backend"
version = "0.1.0"
description = "DonnaAI FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
    "pydantic-settings>=2.10.0,<3.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "alembic>=1.16.0,<2.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "PyJWT>=2.10.0,<3.0.0",
    "httpx>=0.28.0,<1.0.0",
    "redis>=5.0.0,<6.0.0",
    "celery[redis]>=5.4.0,<6.0.0",
    "python-socketio>=5.12.0,<6.0.0",
    "openai>=2.26.0,<3.0.0",
    "langgraph>=1.1.0,<2.0.0",
    "langchain-core>=1.2.0,<2.0.0",
    "langchain-openai>=1.1.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0,<9.0.0",
    "httpx>=0.28.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
